+++ src/app/scans/WebApi/hooks/useInterval.ts
@@ -1,0 +1,22 @@
+// hooks/useInterval.ts
+import { useEffect, useRef } from "react";
+
+export const useInterval = (
+  callback: () => void,
+  delay: number,
+  shouldRun: boolean
+) => {
+  const savedCallback = useRef<(() => void) | null>(null);
+
+  useEffect(() => {
+    savedCallback.current = callback;
+  }, [callback]);
+
+  useEffect(() => {
+    if (!shouldRun || !delay) return;
+    const id = setInterval(() => {
+      savedCallback.current?.();
+    }, delay);
+    return () => clearInterval(id);
+  }, [delay, shouldRun]);
+};
