+++ src/app/scans/WebApi/types/scanTypes.ts
@@ -1,0 +1,8 @@
+export interface Scan {
+  reqCount: string;
+  alertCount: string;
+  progress: string;
+  newAlertCount: string;
+  id: string;
+  state: string;
+}
