+++ src/components/PastScans.tsx
@@ -1,4 +1,4 @@
-// 'use client';
+'use client';
 
 import {
   RiArrowRightUpLine,
